+++ tools/pylint-extensions/azure-pylint-guidelines-checker/tests/test_files/missing_dep_test_pkg/acceptable.py
@@ -1,0 +1,14 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+
+# Acceptable: typing_extensions is in setup.py install_requires
+import typing_extensions
+
+# Acceptable: requests is in setup.py install_requires
+import requests
+
+# Acceptable: stdlib imports are fine (not in _KNOWN_THIRD_PARTY)
+import os
+import json