+++ tools/pylint-extensions/azure-pylint-guidelines-checker/tests/test_files/missing_dep_violation_pkg/violation.py
@@ -1,0 +1,10 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+
+# Violation: typing_extensions is NOT in setup.py install_requires
+import typing_extensions
+
+# Violation: aiohttp is NOT in setup.py install_requires
+from aiohttp import ClientSession